// src/theme/enums.ts
export enum APP_COLORS {
  ROOT_PRIMARY_COLOR = "#FFC107",
  ROOT_PRIMARY_COLOR_LITE = "#FFD54F",
  ROOT_PRIMARY_COLOR_DARK = "#FFAB00",
  ROOT_SECONDARY_COLOR = "#1F1F1F",
  ROOT_SECONDARY_COLOR_LITE = "#9E9E9E",
  ROOT_SECONDARY_COLOR_DARK = "#212121",
  ROOT_MILK_WHITE_COLOR = "#FFFFFF",
  ROOT_PRIMARY_TEXT_COLOR = "#111111",
  ROOT_SECONDARY_TEXT_COLOR = "#767676",
  ROOT_TEXT_COLOR_GRAY = "",
  ROOT_TEXT_COLOR_GRAY_LITE = "#F5F5F5",
  ROOT_DIVIDER_COLOR = "#e0e0e0",
  ROOT_BG_LITE_GRAY = "#F5F5F5",
  SOCIAL_GITHUB_COLOR = "#937DC2",
  SOCIAL_FACEBOOK_COLOR = "#25316D",
  SOCIAL_GOOGLE_COLOR = "#db4437",
  ROOT_LIGHT_BLUE = "#E1F5FE",
}
export enum APP_FONTS {
  ROOT_PRIMARY_FONT = "Jost",
}

export enum APP_ICONS {
  CART_ICON = "CART_ICON",
  FILTER_ICON = "FILTER_ICON",
  HOME_ICON = "HOME_ICON",
  SEARCH_ICON = "SEARCH_ICON",
  USER_ICON = "USER_ICON",
  SHIPPING_ICON = "SHIPPING_ICON",
  MINUS_ICON = "MINUS_ICON",
  PLUS_ICON = "PLUS_ICON",
}